Make item and author metadata fetching work
This still blows up way too often due to missing metadata
in many items, but for reasonably complete metadata sets,
it works. More work is still needed to make placeOfBirth
usable, though - currently, it is simply a name of a place
like "Frankfurt am Main", while we actually need to know
if, e.g., the author was born it a TRIPTIS state.